Improved smoothness of FixedFpsMainLoop implementation.

+++ jounce-core/src/main/java/com/thrashplay/jounce/entity/ai/RandomBehaviorPaddleController.java
@@ -37,7 +37,7 @@ public class RandomBehaviorPaddleController implements Updateable {
     }
 
     @Override
-    public void update(long delta) {
+    public void update(float delta) {
         if (System.currentTimeMillis() > behaviorEndTime) {
             changeBehavior();
         }
